+++ xposed/src/main/java/moe/fuqiuluo/shamrock/remote/action/handlers/GetTroopMemberInfo.kt
@@ -23,7 +23,7 @@ internal object GetTroopMemberInfo: IActionHandler() {
         }.getOrThrow()
 
         return ok(SimpleTroopMemberInfo(
-            uin = info.memberuin,
+            uin = info.memberuin.toLong(),
             name = info.friendnick.ifNullOrEmpty(info.autoremark) ?: "",
             showName = info.troopnick.ifNullOrEmpty(info.troopColorNick),
             distance = info.distance,
@@ -31,7 +31,7 @@ internal object GetTroopMemberInfo: IActionHandler() {
             joinTime = info.join_time,
             lastActiveTime = info.last_active_time,
             uniqueName = info.mUniqueTitle,
-            groupId = groupId,
+            groupId = groupId.toLong(),
             nick = info.friendnick.ifNullOrEmpty(info.autoremark) ?: "",
             sex = when(info.sex.toShort()) {
                 Card.FEMALE -> "female"
